+++ torch2trt/contrib/qat/layers/quant_pooling.py
@@ -98,7 +98,7 @@ class QuantMaxPool2d(torch.nn.Module,_utils.QuantInputMixin):
         else:
             quant_input = torch.fake_quantize_per_channel_affine(input,
                     quantizer.quant_scale,
-                    quantizer.zero_point,
+                    quantizer.zero_point.to(torch.long),
                     quantizer.quant_axis.to(torch.int32).item(),
                     quantizer.quant_min.to(torch.long).item(),
                     quantizer.quant_max.to(torch.long).item())
@@ -197,7 +197,7 @@ class QuantAdaptiveAvgPool2d(torch.nn.Module,_utils.QuantInputMixin):
         else:
             quant_input = torch.fake_quantize_per_channel_affine(input,
                     quantizer.quant_scale,
-                    quantizer.zero_point,
+                    quantizer.zero_point.to(torch.long),
                     quantizer.quant_axis.to(torch.int32).item(),
                     quantizer.quant_min.to(torch.long).item(),
                     quantizer.quant_max.to(torch.long).item())
